updating UTCI method and fixing multiple annoyances
