+++ LadybugTools_Engine/Python/src/ladybugtools_toolkit/external_comfort/spatial/spatial_comfort.py
@@ -42,7 +42,8 @@ from ..simulate import SimulationResult
 from ..simulate import direct_sun_hours as dsh
 from ..simulate import working_directory
 from ..utci import describe_monthly as describe_utci_monthly
-from ..utci import utci, utci_parallel
+from ..utci import utci
+from ..utci.calculate import utci_parallel
 from .calculate import (
     rwdi_london_thermal_comfort_category,
     shaded_unshaded_interpolation,
@@ -145,7 +146,7 @@ class SpatialComfort(BHoMObject):
         self._triangulation = create_triangulation(
             self.points.x.values,
             self.points.y.values,
-            alpha=1.1,
+            alpha=3,
             max_iterations=250,
             increment=0.01,
         )
